+++ src/OpenTelemetry.Exporter.Jaeger/Implementation/JaegerActivityExtensions.cs
@@ -50,12 +50,11 @@ namespace OpenTelemetry.Exporter.Jaeger.Implementation
             [SemanticConventions.AttributeDbInstance] = 2, // peer.service for Redis.
         };
 
-        private static readonly ListEnumerator<ActivityEvent, PooledListState<JaegerLog>>.ForEachDelegate ProcessActivityEventRef = ProcessActivityEvent;
         private static readonly DictionaryEnumerator<string, object, PooledListState<JaegerTag>>.ForEachDelegate ProcessTagRef = ProcessTag;
 
         public static JaegerSpan ToJaegerSpan(this Activity activity)
         {
-            var jaegerTags = new TagState
+            var jaegerTags = new TagEnumerationState
             {
                 Tags = PooledList<JaegerTag>.Create(),
             };
@@ -156,33 +155,25 @@ namespace OpenTelemetry.Exporter.Jaeger.Implementation
                 duration: (long)activity.Duration.TotalMilliseconds * 1000,
                 references: activity.ToJaegerSpanRefs(),
                 tags: jaegerTags.Tags,
-                logs: activity.Events.ToJaegerLogs());
+                logs: activity.ToJaegerLogs());
         }
 
         public static PooledList<JaegerSpanRef> ToJaegerSpanRefs(this Activity activity)
         {
-            LinkState references = default;
+            LinkEnumerationState references = default;
 
             activity.EnumerateLinks(ref references);
 
-            return references.List;
+            return references.SpanRefs;
         }
 
-        public static PooledList<JaegerLog> ToJaegerLogs(this IEnumerable<ActivityEvent> events)
+        public static PooledList<JaegerLog> ToJaegerLogs(this Activity activity)
         {
-            PooledListState<JaegerLog> logs = default;
+            EventEnumerationState logs = default;
 
-            if (events == null)
-            {
-                return logs.List;
-            }
+            activity.EnumerateEvents(ref logs);
 
-            ListEnumerator<ActivityEvent, PooledListState<JaegerLog>>.AllocationFreeForEach(
-                events,
-                ref logs,
-                ProcessActivityEventRef);
-
-            return logs.List;
+            return logs.Logs;
         }
 
         public static JaegerLog ToJaegerLog(this ActivityEvent timedEvent)
@@ -284,7 +275,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Implementation
             }
         }
 
-        private static void ProcessJaegerTag(ref TagState state, string key, JaegerTag jaegerTag)
+        private static void ProcessJaegerTag(ref TagEnumerationState state, string key, JaegerTag jaegerTag)
         {
             if (jaegerTag.VStr != null)
             {
@@ -315,18 +306,6 @@ namespace OpenTelemetry.Exporter.Jaeger.Implementation
             PooledList<JaegerTag>.Add(ref state.Tags, jaegerTag);
         }
 
-        private static bool ProcessActivityEvent(ref PooledListState<JaegerLog> state, ActivityEvent e)
-        {
-            if (!state.Created)
-            {
-                state.List = PooledList<JaegerLog>.Create();
-                state.Created = true;
-            }
-
-            PooledList<JaegerLog>.Add(ref state.List, e.ToJaegerLog());
-            return true;
-        }
-
         private static bool ProcessTag(ref PooledListState<JaegerTag> state, KeyValuePair<string, object> attribute)
         {
             if (attribute.Value is Array)
@@ -341,7 +320,7 @@ namespace OpenTelemetry.Exporter.Jaeger.Implementation
             return true;
         }
 
-        private struct TagState : IActivityEnumerator<KeyValuePair<string, object>>
+        private struct TagEnumerationState : IActivityEnumerator<KeyValuePair<string, object>>
         {
             public PooledList<JaegerTag> Tags;
 
@@ -370,21 +349,41 @@ namespace OpenTelemetry.Exporter.Jaeger.Implementation
             }
         }
 
-        private struct LinkState : IActivityEnumerator<ActivityLink>
+        private struct LinkEnumerationState : IActivityEnumerator<ActivityLink>
         {
             public bool Created;
 
-            public PooledList<JaegerSpanRef> List;
+            public PooledList<JaegerSpanRef> SpanRefs;
 
             public bool ForEach(ActivityLink activityLink)
             {
                 if (!this.Created)
                 {
-                    this.List = PooledList<JaegerSpanRef>.Create();
+                    this.SpanRefs = PooledList<JaegerSpanRef>.Create();
                     this.Created = true;
                 }
 
-                PooledList<JaegerSpanRef>.Add(ref this.List, activityLink.ToJaegerSpanRef());
+                PooledList<JaegerSpanRef>.Add(ref this.SpanRefs, activityLink.ToJaegerSpanRef());
+
+                return true;
+            }
+        }
+
+        private struct EventEnumerationState : IActivityEnumerator<ActivityEvent>
+        {
+            public bool Created;
+
+            public PooledList<JaegerLog> Logs;
+
+            public bool ForEach(ActivityEvent activityEvent)
+            {
+                if (!this.Created)
+                {
+                    this.Logs = PooledList<JaegerLog>.Create();
+                    this.Created = true;
+                }
+
+                PooledList<JaegerLog>.Add(ref this.Logs, activityEvent.ToJaegerLog());
 
                 return true;
             }